pipeline/requirements.txt
geopandas>=0.13.0
rasterio>=1.3.0
fiona>=1.9.0
shapely>=2.0.0
pyproj>=3.4.0
psycopg2-binary>=2.9.0
boto3>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
fastparquet>=2023.0.0
folium>=0.14.0
flask>=2.0.0

